refactor(ui): consolidate track collections into a single CRUD page
Rework the track-collection admin feature into a classic CRUD layout and
collapse the two browse pages into one:

- Single /maestro/track-collections list page with two modes: management
  (Manage + Create buttons) and session-import (?sessionId, Import-to-session
  + track preview)
- Dedicated create (/new) and edit (/:collectionId) pages
- Extract shared styles/helpers into track-collection-shared.tsx
- Add getTrackCollection API client (GET /track-collections/:id)
- Fix track duration display to use durationInMsToString (value is ms)
- Navigate to the API-returned collection id after create/import

Removes the standalone read-only page and the separate management page.
