+++ apps/rpg-maestro-ui/src/app/app.tsx
@@ -17,7 +17,8 @@ import { ThemeProvider } from '@mui/material/styles';
 import { AdminBoard } from './admin-ui/admin-board';
 import { LoginPage } from './auth/LoginPage';
 import { TrackCollections } from './maestro-ui/track-collections/track-collections';
-import { TrackCollectionsManagement } from './maestro-ui/track-collections/track-collections-management';
+import { TrackCollectionCreate } from './maestro-ui/track-collections/track-collection-create';
+import { TrackCollectionEdit } from './maestro-ui/track-collections/track-collection-edit';
 import { useEffect, useLayoutEffect, useState } from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 import { initAuthRequirements } from './utils/authenticated-fetch';
@@ -58,7 +59,8 @@ export function App() {
             <Route path="/maestro/:sessionId" element={<MaestroSoundboard />} />
             <Route path="/maestro/manage/:sessionId" element={<TracksManagement />} />
             <Route path="/maestro/track-collections" element={<TrackCollections />} />
-            <Route path="/maestro/track-collections/manage" element={<TrackCollectionsManagement />} />
+            <Route path="/maestro/track-collections/new" element={<TrackCollectionCreate />} />
+            <Route path="/maestro/track-collections/:collectionId" element={<TrackCollectionEdit />} />
             <Route path="/maestro/admin" element={<AdminBoard />} />
             <Route path="/admin" element={<Navigate to="/maestro" replace />} />
           </Routes>
