+++ apps/rpg-maestro-ui/src/app/maestro-ui/track-collections/track-collections.tsx
@@ -1,29 +1,27 @@
 import { useEffect, useState } from 'react';
 import { CollectionTrack, TrackCollection } from '@rpg-maestro/rpg-maestro-api-contract';
 import { withAuthenticationRequired } from '@auth0/auth0-react';
+import { useNavigate, useSearchParams } from 'react-router-dom';
 import KeyboardReturnIcon from '@mui/icons-material/KeyboardReturn';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import ExpandLessIcon from '@mui/icons-material/ExpandLess';
 import MusicNoteIcon from '@mui/icons-material/MusicNote';
-import CollectionsBookmarkTwoTone from '@mui/icons-material/CollectionsBookmarkTwoTone';
+import AddIcon from '@mui/icons-material/Add';
+import SettingsIcon from '@mui/icons-material/Settings';
+import Button from '@mui/material/Button';
+import IconButton from '@mui/material/IconButton';
 
 import { getAllTrackCollections, importCollectionToSession } from '../maestro-api';
 import { Loading } from '../../auth/Loading';
 import { TextLinkWithIconWrapper } from '../../ui-components/text-link-with-icon-wrapper';
 import { isDevModeEnabled } from '../../../FeaturesConfiguration';
-import { formatTodayDate } from '../../utils/time';
-import { useSearchParams } from 'react-router-dom';
+import { durationInMsToString, formatTodayDate } from '../../utils/time';
 import { toast, ToastContainer } from 'react-toastify';
-
-function formatDuration(seconds: number): string {
-  const m = Math.floor(seconds / 60);
-  const s = seconds % 60;
-  return `${m}:${s.toString().padStart(2, '0')}`;
-}
+import { cardStyle, pageStyle, TRACK_COLLECTIONS_ROUTE } from './track-collection-shared';
 
 function TrackPreviewList({ tracks }: { tracks: CollectionTrack[] }) {
   return (
-    <div style={{ paddingTop: '0.5rem', borderTop: '1px solid rgba(255,255,255,0.1)' }}>
+    <div style={{ marginTop: '0.5rem', paddingTop: '0.5rem', borderTop: '1px solid rgba(255,255,255,0.1)' }}>
       {tracks.map((track) => (
         <div
           key={track.id}
@@ -39,7 +37,7 @@ function TrackPreviewList({ tracks }: { tracks: CollectionTrack[] }) {
           <span style={{ flex: 1, fontSize: '0.9rem' }}>{track.name}</span>
           {track.tags.length > 0 && <span style={{ fontSize: '0.75rem', opacity: 0.6 }}>{track.tags.join(', ')}</span>}
           <span style={{ fontSize: '0.8rem', opacity: 0.6, whiteSpace: 'nowrap' }}>
-            {formatDuration(track.duration)}
+            {durationInMsToString(track.duration)}
           </span>
         </div>
       ))}
@@ -53,6 +51,8 @@ type TrackCollectionsContentProps = {
   errorMessage: string | null;
   sessionId?: string;
   onImport?: (collectionId: string) => Promise<void>;
+  onManage?: (collectionId: string) => void;
+  onCreateNew?: () => void;
 };
 
 export function TrackCollectionsContent({
@@ -61,6 +61,8 @@ export function TrackCollectionsContent({
   errorMessage,
   sessionId,
   onImport,
+  onManage,
+  onCreateNew,
 }: TrackCollectionsContentProps) {
   const [importingIds, setImportingIds] = useState<Set<string>>(new Set());
   const [expandedIds, setExpandedIds] = useState<Set<string>>(new Set());
@@ -98,31 +100,22 @@ export function TrackCollectionsContent({
   const backText = sessionId ? 'Back to session' : 'Back to admin';
 
   return (
-    <div
-      style={{
-        height: '100%',
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'flex-start',
-        gap: '1rem',
-        padding: '1rem',
-      }}
-    >
+    <div style={pageStyle}>
       <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between', alignItems: 'center' }}>
         <TextLinkWithIconWrapper link={backLink} text={backText} materialUiIcon={KeyboardReturnIcon} />
         <h1 style={{ margin: 0 }}>Track collections</h1>
-        {!sessionId && (
-          <TextLinkWithIconWrapper
-            link="/maestro/track-collections/manage"
-            text="Manage collections"
-            materialUiIcon={CollectionsBookmarkTwoTone}
-          />
-        )}
       </div>
-      <hr style={{ width: '100vw', borderColor: 'var(--gold-color)' }} />
+      <hr style={{ width: '100%', borderColor: 'var(--gold-color)' }} />
+      {!sessionId && (
+        <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
+          <Button variant="contained" startIcon={<AddIcon />} onClick={() => onCreateNew?.()}>
+            Create new collection
+          </Button>
+        </div>
+      )}
       {isLoading && <p>Loading track collections...</p>}
       {!isLoading && errorMessage && <p>{errorMessage}</p>}
-      {!isLoading && !errorMessage && trackCollections.length === 0 && <p>No track collections found.</p>}
+      {!isLoading && !errorMessage && trackCollections.length === 0 && <p>No track collections yet.</p>}
       {!isLoading && !errorMessage && (
         <div style={{ display: 'flex', flexDirection: 'column', gap: '0.75rem' }}>
           {[...trackCollections]
@@ -131,17 +124,9 @@ export function TrackCollectionsContent({
               const isExpanded = expandedIds.has(collection.id);
               const trackCount = collection.tracks.length;
               return (
-                <div
-                  key={collection.id}
-                  style={{
-                    border: '1px solid rgba(255,255,255,0.15)',
-                    borderRadius: '6px',
-                    padding: '0.75rem 1rem',
-                    background: 'rgba(255,255,255,0.03)',
-                  }}
-                >
+                <div key={collection.id} style={cardStyle}>
                   <div style={{ display: 'flex', alignItems: 'center', gap: '1rem' }}>
-                    <div style={{ flex: 1 }}>
+                    <div style={{ flex: 1, minWidth: 0 }}>
                       <span style={{ fontWeight: 'bold', fontSize: '1.05rem' }}>{collection.name}</span>
                       {collection.description && (
                         <span style={{ marginLeft: '0.75rem', opacity: 0.7, fontSize: '0.9rem' }}>
@@ -155,29 +140,33 @@ export function TrackCollectionsContent({
                     <span style={{ opacity: 0.5, fontSize: '0.8rem', whiteSpace: 'nowrap' }}>
                       {formatTodayDate(collection.updated_at)}
                     </span>
-                    {sessionId && (
-                      <button
+                    {sessionId ? (
+                      <Button
+                        variant="outlined"
                         onClick={() => handleImport(collection.id)}
                         disabled={importingIds.has(collection.id)}
-                        style={{ whiteSpace: 'nowrap' }}
+                        sx={{ whiteSpace: 'nowrap' }}
                       >
                         {importingIds.has(collection.id) ? 'Importing...' : 'Import to session'}
-                      </button>
+                      </Button>
+                    ) : (
+                      <Button
+                        variant="outlined"
+                        startIcon={<SettingsIcon />}
+                        onClick={() => onManage?.(collection.id)}
+                        sx={{ whiteSpace: 'nowrap' }}
+                      >
+                        Manage
+                      </Button>
                     )}
                     {trackCount > 0 && (
-                      <button
+                      <IconButton
+                        size="small"
                         onClick={() => toggleExpanded(collection.id)}
-                        style={{
-                          background: 'none',
-                          border: 'none',
-                          cursor: 'pointer',
-                          padding: '0.25rem',
-                          color: 'inherit',
-                        }}
                         aria-label={isExpanded ? 'Collapse tracks' : 'Expand tracks'}
                       >
                         {isExpanded ? <ExpandLessIcon /> : <ExpandMoreIcon />}
-                      </button>
+                      </IconButton>
                     )}
                   </div>
                   {isExpanded && <TrackPreviewList tracks={collection.tracks} />}
@@ -192,6 +181,7 @@ export function TrackCollectionsContent({
 }
 
 export function TrackCollectionsComponent() {
+  const navigate = useNavigate();
   const [trackCollections, setTrackCollections] = useState<TrackCollection[]>([]);
   const [isLoading, setIsLoading] = useState(true);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
@@ -236,6 +226,8 @@ export function TrackCollectionsComponent() {
       errorMessage={errorMessage}
       sessionId={sessionId}
       onImport={sessionId ? handleImport : undefined}
+      onManage={(id) => navigate(`${TRACK_COLLECTIONS_ROUTE}/${id}`)}
+      onCreateNew={() => navigate(`${TRACK_COLLECTIONS_ROUTE}/new`)}
     />
   );
 }
